+++ plugins/link/lang/cs.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'cs', {
 	anchorName: 'Podle jména kotvy',
 	charset: 'Přiřazená znaková sada',
 	cssClasses: 'Třída stylu',
+	download: 'Force Download', // MISSING
 	emailAddress: 'E-mailová adresa',
 	emailBody: 'Tělo zprávy',
 	emailSubject: 'Předmět zprávy',