+++ plugins/link/lang/de.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'de', {
 	anchorName: 'Nach Ankername',
 	charset: 'Verknüpfter Ressourcenzeichensatz',
 	cssClasses: 'Formatvorlagenklasse',
+	download: 'Download erzwingen',
 	emailAddress: 'E-Mail-Adresse',
 	emailBody: 'Nachrichtentext',
 	emailSubject: 'Betreffzeile',