+++ plugins/link/lang/lv.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'lv', {
 	anchorName: 'Pēc iezīmes nosaukuma',
 	charset: 'Pievienotā resursa kodējums',
 	cssClasses: 'Stilu saraksta klases',
+	download: 'Force Download', // MISSING
 	emailAddress: 'E-pasta adrese',
 	emailBody: 'Ziņas saturs',
 	emailSubject: 'Ziņas tēma',