+++ plugins/link/lang/mn.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'mn', {
 	anchorName: 'Зангуугийн нэрээр',
 	charset: 'Тэмдэгт оноох нөөцөд холбогдсон',
 	cssClasses: 'Stylesheet классууд',
+	download: 'Force Download', // MISSING
 	emailAddress: 'Э-шуудангийн хаяг',
 	emailBody: 'Зурвасны их бие',
 	emailSubject: 'Зурвасны гарчиг',