+++ plugins/link/lang/ru.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'ru', {
 	anchorName: 'По имени',
 	charset: 'Кодировка ресурса',
 	cssClasses: 'Классы CSS',
+	download: 'Force Download', // MISSING
 	emailAddress: 'Email адрес',
 	emailBody: 'Текст сообщения',
 	emailSubject: 'Тема сообщения',