+++ plugins/link/lang/sv.js
@@ -19,6 +19,7 @@ CKEDITOR.plugins.setLang( 'link', 'sv', {
 	anchorName: 'Efter ankarnamn',
 	charset: 'Teckenuppställning',
 	cssClasses: 'Stilmall',
+	download: 'Force Download', // MISSING
 	emailAddress: 'E-postadress',
 	emailBody: 'Innehåll',
 	emailSubject: 'Ämne',